requirements.txt
Django==4.2.5
django-environ==0.11.2
psycopg2==2.9.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-taggit==4.0.0
